page based routing added for aws deployement
