+++ .env
@@ -1,2 +1,2 @@
-NEXT_PUBLIC_API_BASE_URL=http://localhost:3001
+NEXT_PUBLIC_API_BASE_URL=https://jira-backend-7e3x.onrender.com
 NODE_ENV=dev
